move prebuild related workflows to their own file
